+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(longurl)
-
-test_check("longurl")